Refactor SearchController and exract Helper
- Reduce the level of indentation by 1
- Encapsulate controller methods in helper module
